pyproject.toml
[project]
name = "codeai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai==1.12.1",
    "openai>=2.11.0",
    "python-dotenv==1.1.0",
    "rich>=14.2.0",
]
